repository

+++ src/main/java/com/example/firstproject/FirstprojectApplication.java
@@ -7,7 +7,7 @@ public class FirstprojectApplication {
 
 	public static void main(String[] args) {
 		int a = 3;
-		int b = 4;
+		double b = 4;
 		SpringApplication.run(Firstgit projectApplication.class, args);
 	}
